CU-86bzg2vxh - Remove validate/matching
