+++ JeMPI_Apps/JeMPI_UI/src/pages/settings/Settings.tsx
@@ -1,51 +1,57 @@
-import { useEffect, useState } from 'react';
-import { Grid, Tab, Tabs, Typography } from '@mui/material';
-import { Box } from '@mui/system';
-import { SyntheticEvent } from 'react';
-import CommonSettings from './common/Common';
-import UniqueToGR from './uniqueToGR/UniqueToGR';
-import UniqueToInteraction from './uniqueToInteraction/UniqueToInteraction';
-import Deterministic from './deterministic/Deterministic';
-import Blocking from './blocking/Blocking';
-import GoldenRecordLists from './goldenRecordLists/GoldenRecordLists';
-import InteractiveNode from './interactiveNode/InteractiveNode';
-import { CustomTabPanel, a11yProps } from './deterministic/BasicTabs';
-import { Configuration } from 'types/Configuration';
-import { generateId } from 'utils/helpers';
-import Probabilisitic from './probabilistic/Probabilistic';
+import { useEffect, useState } from 'react'
+import { Grid, Tab, Tabs, Typography } from '@mui/material'
+import { Box } from '@mui/system'
+import { SyntheticEvent } from 'react'
+import CommonSettings from './common/Common'
+import UniqueToGR from './uniqueToGR/UniqueToGR'
+import UniqueToInteraction from './uniqueToInteraction/UniqueToInteraction'
+import Deterministic from './deterministic/Deterministic'
+import Blocking from './blocking/Blocking'
+import GoldenRecordLists from './goldenRecordLists/GoldenRecordLists'
+import InteractiveNode from './interactiveNode/InteractiveNode'
+import { CustomTabPanel, a11yProps } from './deterministic/BasicTabs'
+import { Configuration } from 'types/Configuration'
+import { generateId } from 'utils/helpers'
+import Probabilistic from './probabilistic/Probabilistic'
 
 const Settings = () => {
-  const [value, setValue] = useState(0);
+  const [value, setValue] = useState(0)
   const [configurationData, setConfigurationData] = useState(() => {
-    const storedData = localStorage.getItem('configuration');
-    return storedData ? generateId(JSON.parse(storedData)) : ({} as Configuration);
-  });
+    const storedData = localStorage.getItem('configuration')
+    return storedData
+      ? generateId(JSON.parse(storedData))
+      : ({} as Configuration)
+  })
 
   const handleChange = (event: SyntheticEvent, newValue: number) => {
-    setValue(newValue);
-  };
+    setValue(newValue)
+  }
 
   useEffect(() => {
     const handleStorageChange = (event: StorageEvent) => {
       if (event.key === 'configuration') {
-        const newConfigData = localStorage.getItem('configuration');
-        setConfigurationData(newConfigData ? generateId(JSON.parse(newConfigData)) : ({} as Configuration));
+        const newConfigData = localStorage.getItem('configuration')
+        setConfigurationData(
+          newConfigData
+            ? generateId(JSON.parse(newConfigData))
+            : ({} as Configuration)
+        )
       }
-    };
+    }
 
-    window.addEventListener('storage', handleStorageChange);
+    window.addEventListener('storage', handleStorageChange)
 
     return () => {
-      window.removeEventListener('storage', handleStorageChange);
-    };
-  }, [configurationData]);
+      window.removeEventListener('storage', handleStorageChange)
+    }
+  }, [])
 
   useEffect(() => {
-    const storedData = localStorage.getItem('configuration');
+    const storedData = localStorage.getItem('configuration')
     if (storedData) {
-      setConfigurationData(generateId(JSON.parse(storedData)));
+      setConfigurationData(generateId(JSON.parse(storedData)))
     }
-  }, [configurationData]);
+  }, [])
 
   return (
     <Grid container spacing={4}>
@@ -89,14 +95,18 @@ const Settings = () => {
             <Typography variant="h5" sx={{ py: 3 }}>
               Setup common properties
             </Typography>
-            <CommonSettings demographicData={configurationData.demographicFields} />
+            <CommonSettings
+              demographicData={configurationData.demographicFields}
+            />
           </CustomTabPanel>
           <CustomTabPanel value={value} index={1}>
             <Typography variant="h5" sx={{ py: 3 }}>
               Setup properties that are unique to the golden record
             </Typography>
             <UniqueToGR
-              uniqueToGoldenRecordData={configurationData?.auxGoldenRecordFields || []}
+              uniqueToGoldenRecordData={
+                configurationData?.auxGoldenRecordFields || []
+              }
             />
           </CustomTabPanel>
           <CustomTabPanel value={value} index={2}>
@@ -104,7 +114,9 @@ const Settings = () => {
               Setup properties that are unique to the interaction
             </Typography>
             <UniqueToInteraction
-              uniqueInteractionData={configurationData?.auxInteractionFields || []}
+              uniqueInteractionData={
+                configurationData?.auxInteractionFields || []
+              }
             />
           </CustomTabPanel>
           <CustomTabPanel value={value} index={3}>
@@ -129,13 +141,13 @@ const Settings = () => {
           </CustomTabPanel>
           <CustomTabPanel value={value} index={6}>
             <Typography variant="h5" sx={{ py: 3 }}>
-              <Probabilisitic rules={configurationData?.rules || {}}/>
+              <Probabilistic />
             </Typography>
           </CustomTabPanel>
         </Box>
       </Grid>
     </Grid>
-  );
-};
+  )
+}
 
-export default Settings;
+export default Settings
