+++ JeMPI_Apps/JeMPI_UI/src/pages/settings/probabilistic/Probabilistic.tsx
@@ -1,83 +1,334 @@
-import { useEffect, useState } from 'react'
-import { Card, CardContent, Tab, Tabs } from '@mui/material'
+import { Typography, Grid, TextField, Slider } from '@mui/material'
 import { Box } from '@mui/system'
-import { a11yProps, CustomTabPanel } from '../deterministic/BasicTabs'
-import ProbabilisticContent from './ProbabilisticContent'
-import { Rule } from 'types/Configuration'
+import { Form, Formik } from 'formik'
+import {
+  defaultValues,
+  initializeValues,
+  marks
+} from './ProbabilisticConstants'
+import { useConfiguration } from 'hooks/useUIConfiguration'
+import { Configuration } from 'types/Configuration'
+import { LoadingButton } from '@mui/lab'
+import { useEffect, useState } from 'react'
 
-interface ProbabilisticProps {
-  rules: {
-    link?: {
-      probabilistic?: Rule[]
-    }
-    validate?: {
-      probabilistic?: Rule[]
-    }
-    matchNotification?: {
-      probabilistic?: Rule[]
-    }
-  }
+interface Rule {
+  vars: string[]
+  text: string
+  linkThreshold?: number
+  marginWindowSize?: number
+  reviewThresholdRange?: { low: number; high: number }
 }
 
-const Probabilistic = ({ rules = {} }: ProbabilisticProps) => {
-  const [value, setValue] = useState(0)
+const Probabilistic = () => {
+  const { configuration, setConfiguration } = useConfiguration()
+  const [rule, setRule] = useState<Rule>({
+    vars: [],
+    text: '',
+    linkThreshold: defaultValues.linkThreshold,
+    marginWindowSize: defaultValues.marginWindowSize,
+    reviewThresholdRange: {
+      low: defaultValues.minReviewThreshold,
+      high: defaultValues.maxReviewThreshold
+    }
+  })
 
-  const handleChange = (event: React.SyntheticEvent, newValue: number) => {
-    setValue(newValue)
+  useEffect(() => {
+    if (configuration) {
+      const probabilisticRules = configuration.rules.link.probabilistic || []
+      setRule(
+        probabilisticRules[0] || {
+          vars: [],
+          text: '',
+          linkThreshold: defaultValues.linkThreshold,
+          marginWindowSize: defaultValues.marginWindowSize,
+          reviewThresholdRange: {
+            low: defaultValues.minReviewThreshold,
+            high: defaultValues.maxReviewThreshold
+          }
+        }
+      )
+    }
+  }, [configuration])
+
+  const handleUpdateConfiguration = (values: any) => {
+    if (!configuration) return
+
+    const updatedConfiguration: Configuration = {
+      ...configuration,
+      rules: {
+        ...configuration.rules,
+        link: {
+          ...configuration.rules.link,
+          probabilistic: [
+            {
+              ...rule,
+              linkThreshold: parseFloat(values.linkThreshold),
+              reviewThresholdRange: {
+                low: parseFloat(values.minReviewThreshold),
+                high: parseFloat(values.maxReviewThreshold)
+              },
+              marginWindowSize: parseFloat(values.marginWindowSize)
+            }
+          ]
+        }
+      }
+    }
+
+    setConfiguration(updatedConfiguration)
+    localStorage.setItem('configuration', JSON.stringify(updatedConfiguration))
   }
 
+  const initialValues = initializeValues(rule)
+
   return (
-    <Card sx={{ minWidth: 275 }}>
-      <CardContent
-        sx={{
-          width: '100%',
-          display: 'flex',
-          flexDirection: 'column',
-          backgroundColor: '#f5f5f5'
-        }}
-      >
-        <Box sx={{ width: '100%', backgroundColor: '#f5f5f5' }}>
-          <Box>
-            <Tabs
-              value={value}
-              onChange={handleChange}
-              aria-label="basic tabs example"
-            >
-              <Tab label="Linking" {...a11yProps(0)} />
-              <Tab label="Validate" {...a11yProps(1)} />
-              <Tab label="Matching" {...a11yProps(2)} />
-            </Tabs>
+    <Formik
+      enableReinitialize
+      initialValues={initialValues}
+      onSubmit={async values => {
+        const finalValues = {
+          minReviewThreshold:
+            parseFloat(values.minReviewThreshold.toString()) ||
+            defaultValues.minReviewThreshold,
+          linkThreshold:
+            parseFloat(values.linkThreshold.toString()) ||
+            defaultValues.linkThreshold,
+          maxReviewThreshold:
+            parseFloat(values.maxReviewThreshold.toString()) ||
+            defaultValues.maxReviewThreshold,
+          marginWindowSize:
+            parseFloat(values.marginWindowSize.toString()) ||
+            defaultValues.marginWindowSize
+        }
+
+        handleUpdateConfiguration(finalValues)
+      }}
+    >
+      {({
+        values,
+        handleChange,
+        setFieldValue,
+        handleSubmit,
+        isSubmitting,
+        touched,
+        errors
+      }) => (
+        <Box
+          sx={{
+            height: '55vh',
+            width: '100%',
+            display: 'flex',
+            flexDirection: 'column',
+            alignItems: 'center',
+            padding: '20px'
+          }}
+        >
+          <Box
+            sx={{
+              display: 'flex',
+              flexDirection: 'column',
+              width: '100%',
+              padding: '20px'
+            }}
+          >
+        
+            <Form onSubmit={handleSubmit}>
+              <Grid
+                container
+                spacing={2}
+                sx={{ marginBottom: '80px', alignItems: 'center' }}
+              >
+                <Slider
+                  id="slider-summary"
+                  valueLabelDisplay="auto"
+                  step={0.05}
+                  marks={marks}
+                  min={0}
+                  max={1}
+                  value={[
+                    parseFloat(values.minReviewThreshold.toString()),
+                    parseFloat(values.linkThreshold.toString()),
+                    parseFloat(values.maxReviewThreshold.toString())
+                  ]}
+                  onChange={(e, value) => {
+                    if (Array.isArray(value)) {
+                      setFieldValue('minReviewThreshold', value[0].toString())
+                      setFieldValue('linkThreshold', value[1].toString())
+                      setFieldValue('maxReviewThreshold', value[2].toString())
+                    } else {
+                      setFieldValue('linkThreshold', value.toString())
+                    }
+                  }}
+                  sx={{
+                    width: '50%',
+                    '& .MuiSlider-thumb': {
+                      "&[data-index='0']": { backgroundColor: 'red' },
+                      "&[data-index='1']": { backgroundColor: 'green' },
+                      "&[data-index='2']": { backgroundColor: 'blue' }
+                    },
+                    marginBottom: '80px'
+                  }}
+                />
+
+                <Grid item xs={8}>
+                  <Grid container spacing={2} sx={{ alignItems: 'center' }}>
+                    <Grid item xs={4}>
+                      <Typography sx={{ color: 'green' }}>
+                        Link Threshold
+                      </Typography>
+                    </Grid>
+                    <Grid item xs={4}>
+                      <TextField
+                        id="link-threshold"
+                        name="linkThreshold"
+                        label="Enter floating point"
+                        variant="outlined"
+                        size="small"
+                        value={values.linkThreshold}
+                        onChange={handleChange}
+                        error={
+                          touched.linkThreshold && Boolean(errors.linkThreshold)
+                        }
+                        helperText={
+                          touched.linkThreshold && errors.linkThreshold
+                        }
+                        inputProps={{ min: 0.2, max: 0.95, step: 0.01 }}
+                        InputLabelProps={{ style: { color: 'green' } }}
+                      />
+                    </Grid>
+                  </Grid>
+                  <Grid
+                    container
+                    spacing={2}
+                    sx={{ alignItems: 'center', marginTop: '10px' }}
+                  >
+                    <Grid item xs={4}>
+                      <Typography sx={{ color: 'red' }}>
+                        Review Threshold range
+                      </Typography>
+                    </Grid>
+                    <Grid item xs={4}>
+                      <TextField
+                        id="min-review-threshold"
+                        name="minReviewThreshold"
+                        label="Min"
+                        variant="outlined"
+                        size="small"
+                        value={values.minReviewThreshold}
+                        onChange={e => {
+                          if (
+                            +e.target.value <
+                            parseFloat(values.linkThreshold.toString())
+                          ) {
+                            handleChange(e)
+                          }
+                        }}
+                        error={
+                          touched.minReviewThreshold &&
+                          Boolean(errors.minReviewThreshold)
+                        }
+                        helperText={
+                          touched.minReviewThreshold &&
+                          errors.minReviewThreshold
+                        }
+                        inputProps={{
+                          min: 0.19,
+                          max: values.linkThreshold,
+                          step: 0.01
+                        }}
+                        InputLabelProps={{ style: { color: 'red' } }}
+                      />
+                    </Grid>
+                  </Grid>
+                  <Grid
+                    container
+                    spacing={2}
+                    sx={{ alignItems: 'center', marginTop: '10px' }}
+                  >
+                    <Grid item xs={4}>
+                      <Typography sx={{ color: 'blue' }}>
+                        Max Review Threshold
+                      </Typography>
+                    </Grid>
+                    <Grid item xs={4}>
+                      <TextField
+                        id="max-review-threshold"
+                        name="maxReviewThreshold"
+                        label="max"
+                        variant="outlined"
+                        size="small"
+                        value={values.maxReviewThreshold}
+                        onChange={e => {
+                          if (
+                            +e.target.value >
+                            parseFloat(values.linkThreshold.toString())
+                          ) {
+                            handleChange(e)
+                          }
+                        }}
+                        inputProps={{
+                          min: values.linkThreshold,
+                          max: 0.96,
+                          step: 0.01
+                        }}
+                        InputLabelProps={{ style: { color: '#1976D2' } }}
+                        error={
+                          touched.maxReviewThreshold &&
+                          Boolean(errors.maxReviewThreshold)
+                        }
+                        helperText={
+                          touched.maxReviewThreshold &&
+                          errors.maxReviewThreshold
+                        }
+                      />
+                    </Grid>
+                  </Grid>
+                  <Grid
+                    container
+                    spacing={2}
+                    sx={{ alignItems: 'center', marginTop: '10px' }}
+                  >
+                    <Grid item xs={4}>
+                      <Typography>Margin Window Size</Typography>
+                    </Grid>
+                    <Grid item xs={4}>
+                      <TextField
+                        id="margin-window-size"
+                        name="marginWindowSize"
+                        label="Enter floating point"
+                        variant="outlined"
+                        size="small"
+                        value={values.marginWindowSize}
+                        onChange={handleChange}
+                        error={
+                          touched.marginWindowSize &&
+                          Boolean(errors.marginWindowSize)
+                        }
+                        helperText={
+                          touched.marginWindowSize && errors.marginWindowSize
+                        }
+                        inputProps={{ min: 0, max: 0.2, step: 0.01 }}
+                      />
+                    </Grid>
+                  </Grid>
+                </Grid>
+              </Grid>
+
+              <Box sx={{ marginTop: '20px' }}>
+                <LoadingButton
+                  type="submit"
+                  variant="contained"
+                  color="primary"
+                  loadingPosition="start"
+                  disabled={isSubmitting}
+                >
+                  Save
+                </LoadingButton>
+              </Box>
+            </Form>
           </Box>
-          <CustomTabPanel value={value} index={0}>
-            <ProbabilisticContent
-              linkingRules={{
-                link: { probabilistic: rules.link?.probabilistic ?? [] }
-              }}
-              currentTab= {'link'}
-            />
-          </CustomTabPanel>
-          <CustomTabPanel value={value} index={1}>
-            <ProbabilisticContent
-              linkingRules={{
-                validate: { probabilistic: rules.validate?.probabilistic ?? [] }
-              }}
-              currentTab= {'validate'}
-            />
-          </CustomTabPanel>
-          <CustomTabPanel value={value} index={2}>
-            <ProbabilisticContent
-              linkingRules={{
-                matchNotification: {
-                  probabilistic: rules.matchNotification?.probabilistic ?? []
-                }
-                
-              }}
-              currentTab= {'matchNotification'}
-            />
-          </CustomTabPanel>
         </Box>
-      </CardContent>
-    </Card>
+      )}
+    </Formik>
   )
 }
